+++ apps/docs/examples/Preview.ts
@@ -116,6 +116,36 @@ export const Previews: Record<string, Preview> = {
     "SegmentedControl/docs/controlled": {
         component: lazy(() => import("@/../../packages/components/src/SegmentedControl/docs/controlled.tsx"))
     },
+    "Tile/docs/preview": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/preview.tsx"))
+    },
+    "Tile/docs/selected": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/selected.tsx"))
+    },
+    "Tile/docs/disabled": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/disabled.tsx"))
+    },
+    "Tile/docs/disabledSelected": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/disabledSelected.tsx"))
+    },
+    "Tile/docs/TileGroup/preview": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/preview.tsx"))
+    },
+    "Tile/docs/TileGroup/selected": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/selected.tsx"))
+    },
+    "Tile/docs/TileGroup/controlled": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/controlled.tsx"))
+    },
+    "Tile/docs/TileGroup/disabled": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/disabled.tsx"))
+    },
+    "Tile/docs/TileGroup/wrapping": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/wrapping.tsx"))
+    },
+    "Tile/docs/TileGroup/columns": {
+        component: lazy(() => import("@/../../packages/components/src/Tile/docs/TileGroup/columns.tsx"))
+    },
     "ListBox/docs/preview": {
         component: lazy(() => import("@/../../packages/components/src/ListBox/docs/preview.tsx"))
     },
@@ -953,6 +983,9 @@ export const Previews: Record<string, Preview> = {
     "Link/docs/quiet": {
         component: lazy(() => import("@/../../packages/components/src/Link/docs/quiet.tsx"))
     },
+    "Link/docs/standalone": {
+        component: lazy(() => import("@/../../packages/components/src/Link/docs/standalone.tsx"))
+    },
     "Link/docs/reactRouterLink": {
         component: lazy(() => import("@/../../packages/components/src/Link/docs/reactRouterLink.tsx"))
     },
@@ -1168,6 +1201,36 @@ export const Previews: Record<string, Preview> = {
     },
     "Badge/docs/badge/text": {
         component: lazy(() => import("@/../../packages/components/src/Badge/docs/badge/text.tsx"))
+    },
+    "Callout/docs/preview": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/preview.tsx"))
+    },
+    "Callout/docs/variant": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/variant.tsx"))
+    },
+    "Callout/docs/noIcon": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/noIcon.tsx"))
+    },
+    "Callout/docs/customIcon": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/customIcon.tsx"))
+    },
+    "Callout/docs/notDismissable": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/notDismissable.tsx"))
+    },
+    "Callout/docs/inline": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/inline.tsx"))
+    },
+    "Callout/docs/compact": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/compact.tsx"))
+    },
+    "Callout/docs/compactVariant": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/compactVariant.tsx"))
+    },
+    "Callout/docs/compactLink": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/compactLink.tsx"))
+    },
+    "Callout/docs/dismissableCompact": {
+        component: lazy(() => import("@/../../packages/components/src/Callout/docs/dismissableCompact.tsx"))
     },
     "Badge/docs/floatingbadge/preview": {
         component: lazy(() => import("@/../../packages/components/src/Badge/docs/floatingbadge/preview.tsx"))