+++ packages/components/src/SegmentedControl/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./SegmentedControl.tsx";
+export * from "./SegmentedControlContext.ts";
 export * from "./SegmentedControlItem.tsx";
+export * from "./SegmentedControlItemContext.ts";