+++ packages/components/src/Tile/docs/disabled.tsx
@@ -1,0 +1,7 @@
+import { Tile } from "../src/index.ts";
+
+export default function Example() {
+    return (
+        <Tile id="camel" isDisabled>Camel</Tile>
+    );
+}