+++ packages/components/src/Tile/src/TileGroupContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+import type { ContextValue } from "react-aria-components";
+
+import type { TileGroupProps } from "./TileGroup.tsx";
+
+export const TileGroupContext = createContext<ContextValue<TileGroupProps, HTMLDivElement>>({});
+
+TileGroupContext.displayName = "TileGroupContext";