+++ packages/components/src/Tile/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./Tile.tsx";
+export * from "./TileContext.ts";
+export * from "./TileGroup.tsx";
+export * from "./TileGroupContext.ts";